Address some codacy complaints
