+++ ctapipe/instrument/camera/tests/test_readout.py
@@ -1,4 +1,6 @@
 """ Tests for CameraGeometry """
+from itertools import combinations
+
 import numpy as np
 import pytest
 from astropy import units as u
@@ -184,9 +186,38 @@ def test_hashing():
 @pytest.mark.parametrize("name", ["LSTCam", "FlashCam", "NectarCam", "CHEC"])
 def test_camera_from_name(name, svc_path):
     """check we can construct all cameras from name"""
-    import os
-
-    print(os.environ["CTAPIPE_SVC_PATH"], svc_path)
-
     camera = CameraReadout.from_name(name)
     assert str(camera) == name
+
+
+def test_eq(example_subarray):
+    """Test telescopes compare equal / non-equal as expected in example subarray"""
+    for tel1, tel2 in combinations(example_subarray.tel.values(), 2):
+        if tel1.camera_name == tel2.camera_name:
+            assert tel1.camera.readout == tel2.camera.readout
+        else:
+            assert tel1.camera.readout != tel2.camera.readout
+
+
+def test_hash(example_subarray):
+    """Test telescope hashes are equal / non-equal as expected in example subarray"""
+    for tel1, tel2 in combinations(example_subarray.tel.values(), 2):
+        if tel1.camera_name == tel2.camera_name:
+            assert hash(tel1.camera.readout) == hash(tel2.camera.readout)
+        else:
+            assert hash(tel1.camera.readout) != hash(tel2.camera.readout)
+
+
+def test_table_roundtrip(prod5_lst):
+    """Test we can roundtrip using from_table / to_table"""
+    from ctapipe.instrument import CameraReadout
+
+    camera_in = prod5_lst.camera.readout
+
+    table = camera_in.to_table()
+    assert table.meta["TAB_VER"] == CameraReadout.CURRENT_TAB_VERSION
+    assert table.meta["TAB_TYPE"] == "ctapipe.instrument.CameraReadout"
+    camera_out = CameraReadout.from_table(table)
+
+    assert camera_in == camera_out
+    assert hash(camera_in) == hash(camera_out)
